# app/server_fastapi/src/openapi_server/models/order_input.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from sqlite3 import Row
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator, Extra  # noqa: F401


class OrderInput(BaseModel):

    class Config:  # was not included in OpenAPI contract, but necessary
        extra = Extra.forbid

    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    OrderInput - a model defined in OpenAPI

        stoks: The stoks of this OrderInput [Optional].
        quantity: The quantity of this OrderInput [Optional].
    """

    stoks: Optional[str] = Field(alias="stoks", default=None)
    quantity: Optional[float] = Field(alias="quantity", default=None)

    @staticmethod
    def from_dict(db_data: dict) -> OrderInput:
        return OrderInput(stoks=db_data.get('stoks'), quantity=db_data.get('quantity'))

# OrderInput.model_rebuild()